Feature nab 465/update h2 version (#168)
* feat: H2のバージョンを 1.4.200 に更新しました。

* feat: H2のバージョンを 2.1.210 に更新しました。

* feat: JPA簡易検証に補足説明を追記しました。

* feat: DBクリーンアップの際にFKを指定して削除するよう修正しました。

* feat: H2のバージョンを 2.1.214 に更新しました。

* feat: JPA簡易検証に補足説明を修正しました。

* feat: JPA簡易検証の説明を修正しました。

* feat: 表記ゆれ修正

* feat: 最小限の設定で動作するサンプルのH2のバージョンを 2.1.214 に更新しました。

* feat: H2をJPA簡易検証の対象DBから除外しました。

* Revert "feat: H2をJPA簡易検証の対象DBから除外しました。"

This reverts commit a3420b431cd52d98ffe3de6ba4164d3db1c6b71d.

* feat: JPA簡易検証で使用するH2のバージョンを 2.1.214 に更新しました。

* feat: JPA簡易検証の、DBにH2を使用する場合を修正しました。

* feat: BASE TABLE についてのコメントを追記しました。

* style: インデント修正

* テスト実行時のみH2をレガシーモードにするようJPA簡易検証の手順を修正しました。

* JPA簡易検証に失敗する説明を修正しました。

* JPA簡易検証にH2のレガシーモードが使用できるよう、説明を修正しました。

* JPA仕様の説明を一部修正しました。また、H2の互換性の説明を追記しました。

* H2レガシーモードでtestを実施する手順をコマンドプロンプトを使うよう修正しました。

* プロキシ環境下の場合の補足説明を追記しました。

* JPA簡易検証の実施に、事前にmvn installが必要である旨を追記しました。

* JPAのアノテーションがDB製品やDBのバージョンに依存しないよう修正しました。

* レガシーモードについて言及する文章の位置を修正しました。

* 読みやすくするために改行を追加しました。

+++ src/test/resources/jp/co/tis/gsp/tools/dba/mojo/GenerateEntity_test/basic/h2/expected/output/jp/co/tis/gsptest/entity/entity/IndexTest2.java
@@ -36,7 +36,7 @@ public class IndexTest2 implements Serializable {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "INDEX_TEST2_ID", precision = 19, nullable = false, unique = true)
+    @Column(name = "INDEX_TEST2_ID", precision = 64, nullable = false, unique = true)
     public Long getIndexTest2Id() {
         return indexTest2Id;
     }
@@ -54,7 +54,7 @@ public class IndexTest2 implements Serializable {
      *
      * @return SUB_ID_1
      */
-    @Column(name = "SUB_ID_1", precision = 19, nullable = false, unique = false)
+    @Column(name = "SUB_ID_1", precision = 64, nullable = false, unique = false)
     public Long getSubId1() {
         return subId1;
     }
@@ -72,7 +72,7 @@ public class IndexTest2 implements Serializable {
      *
      * @return SUB_ID_2
      */
-    @Column(name = "SUB_ID_2", precision = 19, nullable = false, unique = false)
+    @Column(name = "SUB_ID_2", precision = 64, nullable = false, unique = false)
     public Long getSubId2() {
         return subId2;
     }
